flask-api/requirements.txt
Flask
Pillow
deforum
